refactor(client): rename files from *.d.ts to *.types.ts
