+++ client/src/components/entities/Creators.tsx
@@ -18,7 +18,7 @@
 
 import { People, Person } from "react-bootstrap-icons";
 import { stylesByItemType } from "../../utils/helpers/HelperFunctions";
-import { EntityType } from "./Entities";
+import { EntityType } from "./entities.types";
 
 /**
  *  renku-ui
