+++ client/src/components/entities/Label.tsx
@@ -17,7 +17,7 @@
  */
 
 import { Briefcase, CardList, Diagram3, HddStack } from "react-bootstrap-icons";
-import { EntityType, WorkflowType } from "./Entities";
+import { EntityType, WorkflowType } from "./entities.types";
 
 /**
  *  renku-ui
