+++ client/src/features/display/index.ts
@@ -5,7 +5,7 @@ import {
   toggleSshModal,
   reset,
 } from "./displaySlice";
-import { Display, ProjectConfig } from "./display";
+import { Display, ProjectConfig } from "./display.types";
 
 export { displaySlice, showSshModal, hideSshModal, toggleSshModal, reset };
 export type { Display, ProjectConfig };
