+++ client/src/features/session/components/AboutSessionModal.tsx
@@ -22,7 +22,7 @@ import { Container, Modal, ModalBody, ModalHeader } from "reactstrap";
 
 import { ACCESS_LEVELS } from "../../../api-client";
 import { ExternalLink } from "../../../components/ExternalLinks";
-import { EntityType } from "../../../components/entities/Entities";
+import { EntityType } from "../../../components/entities/entities.types";
 import EntityHeader from "../../../components/entityHeader/EntityHeader";
 import { ProjectMetadata } from "../../../notebooks/components/session.types";
 import { Docs } from "../../../utils/constants/Docs";
