+++ client/src/features/session/components/AnonymousSessionsList.tsx
@@ -23,7 +23,7 @@ import { Col, Row } from "reactstrap";
 import { ErrorAlert, InfoAlert } from "../../../components/Alert";
 import { Loader } from "../../../components/Loader";
 import ContainerWrap from "../../../components/container/ContainerWrap";
-import { User } from "../../../model/RenkuModels";
+import { User } from "../../../model/renkuModels.types";
 import AppContext from "../../../utils/context/appContext";
 import { DEFAULT_APP_PARAMS } from "../../../utils/context/appParams.constants";
 import useLegacySelector from "../../../utils/customHooks/useLegacySelector.hook";
