+++ client/src/project/new/components/Visibility.tsx
@@ -54,7 +54,7 @@ import {
   NewProjectHandlers,
   NewProjectInputs,
   NewProjectMeta,
-} from "./newProject.d";
+} from "./newProject.types";
 
 interface VisibilityProps {
   handlers: NewProjectHandlers;
